CT-58 Membership Reference app extension
