+++ .github/actions/setup-environment/action.yml
@@ -50,21 +50,15 @@ outputs:
   contentful-app-field-as-updated-at:
     description: 'Contentful App Extension Field As Updated At Id'
     value: ${{ steps.vars.outputs.contentful-app-field-as-updated-at }}
-  contentful-app-user-membership:
-    description: 'Contentful App Extension Project Membership'
-    value: ${{ steps.vars.outputs.contentful-app-user-membership }}
   contentful-app-user-positions:
     description: 'Contentful App Extension User Positions Id'
     value: ${{ steps.vars.outputs.contentful-app-user-positions }}
-  contentful-app-user-team-memberships:
-    description: 'Contentful App Extension User Team Memberships Id'
-    value: ${{ steps.vars.outputs.contentful-app-user-team-memberships }}
-  contentful-app-contributing-cohort-memberships:
-    description: 'Contentful App Extension Contributing Cohort Memberships Id'
-    value: ${{ steps.vars.outputs.contentful-app-contributing-cohort-memberships }}
   contentful-app-working-group-deliverables:
     description: 'Contentful App Extension Working Group Deliverables'
     value: ${{ steps.vars.outputs.contentful-app-working-group-deliverables }}
+  contentful-app-membership-reference:
+    description: 'Contentful App Extension Membership Reference'
+    value: ${{ steps.vars.outputs.contentful-app-membership-reference }}
   contentful-environment:
     description: 'Contentful Environment'
     value: ${{ steps.vars.outputs.contentful-environment }}
