+++ .github/environment/Base
@@ -5,16 +5,14 @@ cache-prefix-definitions=definitions-ouput
 cache-prefix-frontend=frontend-output
 cache-prefix-transpile=transpile-output
 contentful-app-clear-on-change=6en82KYFrsxlAIPqypMdOp
-contentful-app-contributing-cohort-memberships=4LchEEroaC90dWNrpTElwg
 contentful-app-date-field-as-first-published-at=6AEbdOZc7KICbwW6KszNlp
 contentful-app-disabled-fields=2finDNk15g5UtOq4DaLNxv
 contentful-app-event-additional-materials=3TfrYX8zVOgFtKxro89UVA
 contentful-app-event-speakers-gp2=3dyTFkiCqTbARDWBzY4HB
 contentful-app-event-speakers-title=6ZkXISzhv1b7jjgyaK2piv
 contentful-app-field-as-updated-at=mqvX9KU5AthRTlnIRhNNh
-contentful-app-user-membership=1k7oc8QBS413LyX8Ed9LnA
+contentful-app-membership-reference=Yp64pYYDuRNHdvAAAJPYa
 contentful-app-user-positions=6gJUoMACdM5WFGRz6j4pbB
-contentful-app-user-team-memberships=2iE2vFoT19Q5u4Rxpio8gz
 contentful-app-working-group-deliverables=2hc2UiWrW8SFhfkKMmGtk9
 contentful-environment=Production
 contentful-organization-id=6Hlb5Pl67A3nWMEyzxnHNZ
