--- packages/contentful-app-extensions/contributing-cohort-memberships/src/react-app-env.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="react-scripts" />
