--- packages/contentful-app-extensions/contributing-cohort-memberships/tsconfig.json
@@ -1,23 +1,0 @@
-{
-  "compilerOptions": {
-    "composite": true,
-    "module": "esnext",
-
-    "allowJs": true,
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "isolatedModules": true,
-    "jsx": "react",
-    "moduleResolution": "node",
-    "noEmit": true,
-    "noFallthroughCasesInSwitch": true,
-    "resolveJsonModule": true,
-    "skipLibCheck": true,
-    "strict": true,
-
-    "target": "es2015",
-    "lib": ["dom", "dom.iterable", "esnext"]
-  },
-  "include": ["src"]
-}
