--- packages/contentful-app-extensions/user-team-memberships/package.json
@@ -1,54 +1,0 @@
-{
-  "name": "@asap-hub/contentful-app-user-team-memberships",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@contentful/app-sdk": "4.22.1",
-    "@contentful/f36-components": "4.48.0",
-    "@contentful/f36-tokens": "4.0.2",
-    "@contentful/field-editor-reference": "^5.9.0",
-    "@contentful/field-editor-shared": "^1.2.0",
-    "@contentful/react-apps-toolkit": "1.2.16",
-    "contentful-management": "10.39.2",
-    "emotion": "10.0.27",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-scripts": "5.0.1"
-  },
-  "scripts": {
-    "start": "cross-env BROWSER=none react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "test:no-watch": "react-scripts test --watchAll=false",
-    "eject": "react-scripts eject",
-    "create-app-definition": "contentful-app-scripts create-app-definition",
-    "upload": "contentful-app-scripts upload --bundle-dir ./build",
-    "upload-ci": "contentful-app-scripts upload --ci --bundle-dir ./build --organization-id $CONTENTFUL_ORG_ID --definition-id $CONTENTFUL_APP_DEF_ID --token $CONTENTFUL_ACCESS_TOKEN"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@contentful/app-scripts": "1.10.2",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "11.2.7",
-    "@tsconfig/create-react-app": "1.0.3",
-    "@types/jest": "29.5.3",
-    "@types/node": "18.17.1",
-    "@types/react": "17.0.62",
-    "@types/react-dom": "17.0.20",
-    "@types/testing-library__jest-dom": "5.14.9",
-    "cross-env": "7.0.3",
-    "typescript": "4.9.5"
-  },
-  "homepage": "."
-}
